use u8

+++ src/engine/gfx/image_manipulation.cpp
@@ -76,10 +76,10 @@ void ConvertToGrayscale(const CImageInfo &Image)
 	const size_t Step = Image.PixelSize();
 	for(size_t i = 0; i < Image.m_Width * Image.m_Height; ++i)
 	{
-		const int R = Image.m_pData[i * Step];
-		const int G = Image.m_pData[i * Step + 1];
-		const int B = Image.m_pData[i * Step + 2];
-		const int Luma = (int)(0.2126f * R + 0.7152f * G + 0.0722f * B);
+		const uint8_t R = Image.m_pData[i * Step];
+		const uint8_t G = Image.m_pData[i * Step + 1];
+		const uint8_t B = Image.m_pData[i * Step + 2];
+		const uint8_t Luma = (uint8_t)(0.2126f * R + 0.7152f * G + 0.0722f * B);
 
 		Image.m_pData[i * Step] = Luma;
 		Image.m_pData[i * Step + 1] = Luma;
